+++ app/models/govuk_publishing_components/component_doc_resolver.rb
@@ -12,8 +12,8 @@ module GovukPublishingComponents
   private
 
     def build(component)
-      fixtures = component[:fixtures].map { |id, data|
-        ComponentFixture.new(id.to_s, data)
+      examples = component[:examples].map { |id, data|
+        ComponentExample.new(id.to_s, data)
       }
 
       ComponentDoc.new(component[:id],
@@ -21,7 +21,7 @@ module GovukPublishingComponents
                        component[:description],
                        component[:body],
                        component[:accessibility_criteria],
-                       fixtures)
+                       examples)
     end
 
     def fetch_component_docs